Remove Lib_PQ_Print in favour of Lib_Print

+++ lib/c/Lib_Print.c
@@ -3,18 +3,18 @@
 #include "Lib_Print.h"
 
 void Lib_Print_print_bytes(uint32_t len, uint8_t* buffer) {
-  for (int i = 0; i < len; i++){
+  for (uint32_t i = 0; i < len; i++){
     printf("%02x ", buffer[i]);
   }
   printf("\n");
 }
 
 void Lib_Print_print_compare(uint32_t len, uint8_t* buffer1, uint8_t* buffer2) {
-  for (int i = 0; i < len; i++){
+  for (uint32_t i = 0; i < len; i++){
     printf("%02x ", buffer1[i]);
   }
   printf("\n");
-  for (int i = 0; i < len; i++){
+  for (uint32_t i = 0; i < len; i++){
     printf("%02x ", buffer2[i]);
   }
   printf("\n");
@@ -22,8 +22,8 @@ void Lib_Print_print_compare(uint32_t len, uint8_t* buffer1, uint8_t* buffer2) {
 
 void Lib_Print_print_compare_display(uint32_t len, uint8_t* buffer1, uint8_t* buffer2) {
   Lib_Print_print_compare(len, buffer1, buffer2);
-  int res = 0;
-  for (int i = 0; i < len; i++) {
+  uint8_t res = 0;
+  for (uint32_t i = 0; i < len; i++) {
     res |= buffer1[i] ^ buffer2[i];
   }
   if (res == 0) {
